+++ cynic-querygen/tests/starwars-tests.rs
@@ -24,3 +24,4 @@ test_query_file!(test_nested_arguments, "nested-arguments.graphql");
 test_query_file!(bare_selection_sets, "bare-selection-set.graphql");
 test_query_file!(multiple_queries, "multiple-queries.graphql");
 test_query_file!(fragment_spreads, "fragment-spreads.graphql");
+test_query_file!(aliases, "aliases.graphql");